requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
mysql-connector-python
sqlalchemy
